changed player action screen to Action.py and added color coding
